refactor(member): migrate tobe to base
