+++ src/main/kotlin/apply/domain/member/Member.kt
@@ -1,46 +1,37 @@
 package apply.domain.member
 
 import support.domain.BaseRootEntity
+import support.infra.PersistenceOnly
 import java.time.LocalDate
 import javax.persistence.AttributeOverride
 import javax.persistence.CascadeType
 import javax.persistence.Column
 import javax.persistence.Embedded
 import javax.persistence.Entity
+import javax.persistence.EnumType
+import javax.persistence.Enumerated
 import javax.persistence.OneToOne
 
 @Entity
 class Member(
     information: MemberInformation,
+
     @AttributeOverride(name = "value", column = Column(name = "password", nullable = false))
     @Embedded
     var password: Password,
+
     authorizationRequirement: AuthorizationRequirement,
-    id: Long = 0L,
-) : BaseRootEntity<Member>(id) {
-    @OneToOne(mappedBy = "member", cascade = [CascadeType.PERSIST, CascadeType.REMOVE], orphanRemoval = true)
-    private var _information: MemberInformation? = information
-    private val information: MemberInformation
-        get() = _information ?: MemberInformation.DELETED
-
-    val email: String
-        get() = information.email
-
-    val name: String
-        get() = information.name
-
-    val birthday: LocalDate
-        get() = information.birthday
-
-    val phoneNumber: String
-        get() = information.phoneNumber
-
-    val githubUsername: String
-        get() = information.githubUsername
+    @Column(nullable = false)
+    @Enumerated(EnumType.STRING)
+    var status: MemberStatus = MemberStatus.ACTIVE,
+) : BaseRootEntity<Member>() {
+    @OneToOne(mappedBy = "member", cascade = [CascadeType.PERSIST], orphanRemoval = true)
+    private var _information: MemberInformation? = null
+    val information: MemberInformation get() = _information ?: throw IllegalStateException("회원 정보가 존재하지 않습니다.")
 
     init {
-        information.member = this
         authorizationRequirement.require(information)
+        attachInformation(information)
     }
 
     fun authenticate(password: Password) {
@@ -54,15 +45,12 @@ class Member(
     ) {
         identify(information.same(name, birthday)) { "사용자 정보가 일치하지 않습니다." }
         this.password = Password(password)
-        registerEvent(PasswordResetEvent(id, name, email, password))
+        registerEvent(PasswordResetEvent(id, name, information.email, password))
     }
 
-    fun changePassword(
-        oldPassword: Password,
-        newPassword: Password,
-    ) {
-        identify(this.password == oldPassword) { "기존 비밀번호가 일치하지 않습니다." }
-        this.password = newPassword
+    fun changePassword(oldPassword: Password, newPassword: Password) {
+        identify(password == oldPassword) { "기존 비밀번호가 일치하지 않습니다." }
+        password = newPassword
     }
 
     fun changePhoneNumber(phoneNumber: String) {
@@ -71,17 +59,26 @@ class Member(
 
     fun withdraw(password: Password) {
         identify(this.password == password) { "사용자 정보가 일치하지 않습니다." }
-        information.member = null
-        _information = null
+        status = MemberStatus.WITHDRAWN
+        detachInformation()
     }
 
-    private fun identify(
-        value: Boolean,
-        lazyMessage: () -> Any = {},
-    ) {
+    private fun identify(value: Boolean, lazyMessage: () -> Any) {
         if (!value) {
             val message = lazyMessage()
             throw UnidentifiedMemberException(message.toString())
         }
+    }
+
+    @PersistenceOnly
+    private fun attachInformation(information: MemberInformation) {
+        _information = information
+        information.member = this
+    }
+
+    @PersistenceOnly
+    private fun detachInformation() {
+        _information?.member = null
+        _information = null
     }
 }
